Simplify the edit operations to work over the splice operation
Now instead of having different edit operations we can generalize over
the splice operation (taken from the VecExt trait in the odds crate).
This allows us to greatly simplify the logic of editing and how the undo
is handled. Most importantly, this allows us to re-implement the paste
semantics in overwrite mode.

Right now the implementation of the splice is rather, simple-minded, but
once we get a splice operation that will actually return data, we will
be able to properly implement it in the future.