Stop treating config files specially in the mock file system
This changes the behavior of the mock filesystem for editing existing
files, previously it would raise a 'file-already-exists' error, now it
will allow overwritting. This was thought useful for a specific behavior
in rex, that is no longer relevant and so now the overwrite behavior
should be used for everything.